modified creat.html in checout app

+++ cart/cart.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from django.conf import settings
 from shop.models import Product
 
@@ -29,7 +30,7 @@ class Cart(object):
             cart[str(product.id)]['product'] = product
 
         for item in cart.values():
-            item['price'] = (item['price'])
+            item['price'] = Decimal(item['price'])
             item['total_price'] = item['price'] * item['quantity']
             yield item
 
@@ -72,4 +73,4 @@ class Cart(object):
         self.save()
 
     def get_total_price(self):
-        return sum(float(item['price']) * item['quantity'] for item in self.cart.values())
+        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
